requirements.txt
pyyaml==6.0.1
requests==2.32.3
pydantic==1.10.19
